+++ src/EMerchantPay/Resources/views/backend/emerchantpay_payment_methods/app.js
@@ -5,6 +5,7 @@
 //      {include file="backend/emerchantpay_payment_methods/view/detail/emerchantpay_direct_form.js"}
 //      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_test_mode_store.js"}
 //      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_wpf_tokenization_store.js"}
+//      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_checkout_bank_codes_store.js"}
 //      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_checkout_transaction_types_store.js"}
 //      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_direct_transaction_types_store.js"}
 //      {include file="backend/emerchantpay_payment_methods/store/emerchantpay_checkout_languages_store.js"}